feat(web): inline agent tree/graph view in agents and project-detail pages
Refactors the graph view from a separate navigation target into an inline
view mode rendered in the same content slot as the grid/list, so the filter
bar (status, label) applies to the tree automatically.

Changes:
- Extract `<scion-agent-tree-view>` shared component from agent-graph.ts
  rendering logic. Accepts `agents` as a property (parent does filtering),
  keeps pan/zoom/hover/collapse state internally, includes its own
  Show-users toggle, and supports an optional `focusId` for deep linking.
- `view-toggle.ts`: make 'graph' a fully normal, clickable mode — remove
  `graphHref` link-out behaviour, always render a button, persist 'graph'
  in localStorage identically to 'grid'/'list'.
- `agent-graph.ts`: slim to a thin wrapper that still fetches its own
  data and owns the cross-project filter, but delegates rendering to
  `<scion-agent-tree-view>`.
- `agents.ts`: remove `graphHref`, render `<scion-agent-tree-view>` inline
  using the same `displayAgents` array the grid/list use; remove now-dead
  `onViewChange` guard; persist 'graph' in stored view mode.
- `project-detail.ts`: same changes as agents.ts.
